+++ sources/afx/video/afx_video_mjpeg+/base64.cpp
@@ -27,16 +27,16 @@ namespace CVSandbox { namespace Video { namespace MJpeg { namespace Private
 static const char basis_64[] = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/";
 
 // Get length of encoded source for the given length of a plain source
-int Base64EncodeLength( int sourceLength )
+size_t Base64EncodeLength( size_t  sourceLength )
 {
     return ( ( sourceLength + 2 ) / 3 * 4 ) + 1;
 }
 
 // Encodes the specifed source using Base64 encoding
-int Base64Encode( char* encoded, const char* source, int sourceLength )
+size_t  Base64Encode( char* encoded, const char* source, size_t  sourceLength )
 {
     char*   ptr = encoded;
-    int     i;
+    size_t  i;
 
     for ( i = 0; i < sourceLength - 2; i += 3 )
     {
